requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
docopt==0.6.2
idna==2.10
Js2Py==0.70
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.19.4
opencv-python==4.5.1.48
packaging==20.8
pandas==1.2.0
Pillow==8.1.0
pipwin==0.5.0
PyAudio @ file:///C:/Users/2kani/pipwin/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pynput==1.7.2
pyparsing==2.4.7
pyperclip==1.8.1
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
tzlocal==2.1
urllib3==1.26.2
